demo/requirements.txt
# 小样本缺陷检测系统依赖
# 适用于 anomalib 0.7.x

# 核心依赖
anomalib==0.7.0
pytorch-lightning>=1.9.0,<2.0.0
torch>=1.13.0
torchvision>=0.14.0

# 数据处理
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0
albumentations>=1.3.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.12.0
tqdm>=4.64.0

# 配置管理
PyYAML>=6.0
omegaconf>=2.3.0

# 工具库
scikit-learn>=1.0.0
scipy>=1.9.0

# 可选：ONNX 导出支持
# onnx>=1.12.0
# onnxruntime>=1.12.0

# 可选：TensorRT 加速
# tensorrt>=8.5.0
